reusable dialog component added and used for data entry delete

+++ src/app/components/common/data-card-list/data-card-list.component.ts
@@ -1,14 +1,16 @@
-import { Component, Injectable, OnInit } from '@angular/core';
+import { Component, Injectable, inject, OnInit } from '@angular/core';
 import { CommonModule, DatePipe } from '@angular/common';
 import { Router } from '@angular/router';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
 import { PageEvent, MatPaginatorIntl, MatPaginatorModule } from '@angular/material/paginator';
+import { MatDialog } from '@angular/material/dialog';
 
 import { ChartDataReadService, ChartDataDeleteService } from 'services';
 import { ChartFormData } from 'interfaces';
 import { DateFormatterPipe, TelephoneFormatterPipe } from 'pipes';
+import { DataAlertDialogComponent } from 'components/common/data-alert-dialog/data-alert-dialog.component';
 
 @Injectable()
 export class PaginatorIntl extends MatPaginatorIntl {
@@ -52,8 +54,11 @@ export class DataCardListComponent implements OnInit {
 
   pageEvent!: PageEvent;
 
+  
+
   constructor(
     private _router: Router,
+    private _dialog: MatDialog,
     private _chartService: ChartDataReadService,
     private _deleteService: ChartDataDeleteService) { }
 
@@ -69,9 +74,26 @@ export class DataCardListComponent implements OnInit {
     this._router.navigate(['/edit-form'], { queryParams: { id: card._id } });
   }
 
-  onCardDelete(card: ChartFormData) {
-    this._deleteService.deleteChartData(card._id).subscribe(() => {
-      this.cardsData = this.cardsData.filter(item => item._id !== card._id);
+  onCardDelete(card: ChartFormData): void {
+    const dialogRef = this._dialog.open(DataAlertDialogComponent, {
+      width: '400px',
+      data: {
+        title: 'Confirm Deletion',
+        message: `Are you sure you want to delete the item "${card.productName}"?`,
+        confirmText: 'Delete',
+        cancelText: 'Cancel'
+      }
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      console.log('Dialog result:', result);
+      if (result) {
+        // User confirmed deletion
+        this._deleteService.deleteChartData(card._id).subscribe(() => {
+          this.cardsData = this.cardsData.filter((item) => item._id !== card._id);
+          this.updatePaginatedCards();
+        });
+      }
     });
   }
